requirements.txt
asgiref==3.7.2
astroid==2.15.8
asttokens==2.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
devtools==0.12.2
dill==0.3.7
Django==4.2.6
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
docutils==0.20.1
drf-yasg==1.21.7
exceptiongroup==1.1.3
executing==2.0.0
Faker==19.11.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
pycparser==2.21
Pygments==2.16.1
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.0
wrapt==1.15.0

coverage==7.3.2
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-dotenv==0.5.2
pytest-mock==3.12.0